MIPS: mips_device.go: Multilib sanitizer support
This commit adds support for linking against mips32r6 sanitizer
libraries which are required when compiling for this architecture.
New set of sanitizer libraries inside clang compiler are tagged
using the "mips32r6" in their name so logic is added to select
appropriate libraries depending on the arch variant being compiled.

Change-Id: I732d9d624f6c3788a826f9eb7902d880c4f99571
Signed-off-by: Miodrag Dinic <miodrag.dinic@mips.com>
Signed-off-by: Lazar Trsic <Lazar.Trsic@mips.com>