MIPS: dumpvars.go: Export ARCH_MIPS_REV6 to the environment
This is required to properly select appropriate directory in clang containing
target static libraries. The issue is that MIPS has 2 32bit architectures
(MIPS32 R < 6 & MIPS32R6) supported which need to have separate sets of libraries
in clang and they are located in "mipsel" and "mips32r6el" directories respectively.

prebuilts/clang/host/linux-x86/soong/clangprebuilts.go contains implementation
for selecting appropriate directories for the target architectures and MIPS
so far contained a hardcoded "mipsel" path. Since there are no information
about the currently compiled target variant in clangprebuilts.go, the only
way to acquire that information is through environment variables.
ARCH_MIPS_REV6 is only set if mips32r6 is compiled as first or second arch
and it is used as indication to select "mips32r6el" directory in clang.

Change-Id: Iea8280977474d2d46d346b53d9d01686fb7bcbb5
Signed-off-by: Miodrag Dinic <miodrag.dinic@mips.com>