MIPS: Override LTO defaults when compiling for mips32r6
When LTO is enabled, the compiler generates LLVM bitcode .o files
which can be linked with native .o files into a library. Before the
actual linking, the code generator needs to generate native code out
of the given bitcode. LLVMgold plugin tells the generator for which
cpu should the code be generated via the -mcpu flag value, which it
optains based on the triple from the -target flag value given to clang.
The default cpu for mipsel-linux-android is mips32, which causes r1
code to be generated, and the linking with other r6 .o files fails,
so pass -mips32r6 to override this behaviour.

Change-Id: Ib02ec71b77d2375352ac17e89cad0b7a37f9e23e

Signed-off-by: Dejan Jovicevic <dejan.jovicevic@mips.com>
Signed-off-by: Lazar Trsic <Lazar.Trsic@mips.com>